// src/app/preparation/equipe/models/equipe.model.ts
export interface Equipe {
    nom: string;
    prix: number;
    capaciteTravail: number;
    capaciteOtage: number;
    capaciteArme: number;
    capaciteColere: number;
    capaciteAngoisse: number;
    capaciteNegociation: number;
    capaciteNouvelleTechnologie: number;
    capaciteExplosif: number;
    etatDeFolie: boolean;
    etatDeSante: string;
    optimiste: number;
    image: string;
    communication: string;

}